# Improve memory access pattern and histogram clipping for CLAHE