While working on #80, I saw some potential for performance improvements in the "original" CLAHE method and made the following changes:
- The **memory access pattern** for updating the sliding window histogram was column-wise on a row-major image and binning computations were done for all updates. I introduced a column-major array of precomputed bin indices, which improved the performance of the histogram updates. This trades off slightly higher memory consumption with faster run time.
- The **histogram clipping method** made quite a lot of passes over the histogram data to redistribute excess values. I implemented a variant of the "look-ahead method" described in a paper titled [Resource efficient real-time processing of Contrast Limited Adaptive Histogram Equalization](https://ieeexplore.ieee.org/document/7577362). This seemed to reduce the number of passes over the data to about 3 for every histogram. In theory, this also affects the "fast" option, but only in
- The **number of bins** that was used internally was more than the number of bins specified. E.g., for 256 user-specified bins (the default) , there were 257 bins used internally, which aligns badly with the value range of uint8 that is used internally.

The changes introduce slight differences in the actual output values, but I verified that they stay within 1 intensity unit in an uint8 image. The following table shows the run times before and after the changes (best of 3 runs). The numbers are `width x height, blockRadius`.

| | Before | After | Speedup |
| - | -----| ----- | ------- |
|100x100, 7| 8ms | 5ms | 1.6 |
|1000x1000, 7| 661ms | 235ms | 2.8 |
|1000x1000, 127| 1329ms | 661ms | 2.0 |
|5000x5000, 127| 36113 | 14292ms | 2.5 |

In principle, the improved histogram clipping also affects the "fast" option. Since the histogram is not computed that often in this case, I saw only improvements of about 20% for small values of `blockRadius` and large images (i.e., many histograms to compute), but no significant speedup for the other cases.

That being said, the ‘fast’ option is still orders of magnitude faster and should definitely be the go-to method unless it’s verifiably insufficient for the use case.

Let me know what you think @axtimwalde!